challenge-2/requirements.txt
flask
requests
python-dotenv
prometheus_client